% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{aaa}
\alias{aaa}
\title{Workaround: Avoid backtraces in examples}
\description{
This example will run first and set the option for the process that builds
the example.
(By default, pkgdown builds in a separate process.)
This also produces a help page that isn't linked from anywhere.
}
\examples{
options(rlang_backtrace_on_error = "none")
}
\keyword{internal}
