% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_chk}
\alias{abort_chk}
\title{Abort Check}
\usage{
abort_chk(...)
}
\arguments{
\item{...}{A substituted object to deparse.}
}
\value{
Throws an error of class \code{chk_error}.
}
\description{
A wrapper on \code{\link[rlang]{abort}(paste0(...))} that throws an error.
}
\details{
It is exported to allow users to construct their own `chk_` functions.
}
\examples{
try(abort_chk())
try(abort_chk("x must be NULL."))
}
\seealso{
\code{\link[rlang]{abort}}
}
