% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Concatenate with Commas}
\usage{
cc(x, con = " or ", bracket = "", chk = TRUE)
}
\arguments{
\item{x}{The object to concatenate.}

\item{con}{A string of the conjunction to separate the last value by.}

\item{bracket}{A string to bracket the values by.}

\item{chk}{A flag specifying whether to check the other arguments.}
}
\value{
A string.
}
\description{
Concatenates object values into a string with each value separated by a comma
and the last value separated by a conjunction.
}
\details{
If x has more than 10 values an ellipsis is used to ensure only 10 values 
are displayed (including the ellipsis)
}
\examples{
cc(1:2)
cc(1:2, con = ",")
cc(3:1, bracket = "'")
cc(1:11)
}
