% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-set.R
\name{check_set}
\alias{check_set}
\title{Check Set}
\usage{
check_set(x, set, exclusive = FALSE, order = FALSE, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{set}{The set to check against.}

\item{exclusive}{A flag specifying whether x must only contain the elements of set.}

\item{order}{A flag specifying whether the order of elements in x must match those in set.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
An informative error if the test fails.
}
\description{
Checks whether the elements of an object match a set.

Useful for checking names.
}
\examples{
check_set(1:2, 2:1)
try(check_set(1:2, 2:1, order = TRUE))
}
\seealso{
Other check: 
\code{\link{check_length}()},
\code{\link{check_values}()}
}
\concept{check}
