% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-package.R
\docType{package}
\name{chk-package}
\alias{chk}
\alias{chk-package}
\title{chk: Check User-Supplied Function Arguments}
\description{
For developers to check user-supplied function arguments.
        It is designed to be simple, fast and customizable.
        Error messages follow the tidyverse style guide.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/chk}
  \item Report bugs at \url{https://github.com/poissonconsulting/chk/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
