% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-all-na.R
\name{chk_all_na}
\alias{chk_all_na}
\alias{vld_all_na}
\title{Check All Missing Values}
\usage{
chk_all_na(x, x_name = NULL)

vld_all_na(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails or
returns the original object if successful so it can used in pipes.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if all missing values using

\code{all(is.na(x))}

\strong{Pass}: \code{NA}, \code{c(NA, NA)}, \code{logical(0)}.

\strong{Fail}: \code{1}, \code{1:2}, \code{"1"}, \code{c(1, NA)}.
}
\section{Functions}{
\itemize{
\item \code{vld_all_na()}: Validate All Missing Values

}}
\examples{
# chk_all_na
try(chk_all_na(1))
try(chk_all_na(c(1, NA)))
chk_all_na(NA)
chk_all_na(c(NA, NA_character_, NA_real_))
# vld_all_na
vld_all_na(1)
vld_all_na(1:2)
vld_all_na(NA_real_)
vld_all_na(integer(0))
vld_all_na(c(NA, 1))
vld_all_na(TRUE)
vld_all_na(c(NA_real_, NA_character_))
}
\seealso{
For more details about the use of this function,
please read the article
\code{vignette("chk-families")}.

Other misc_checkers:
\code{\link[=chk_join]{chk_join()}},
\code{\link[=chk_not_any_na]{chk_not_any_na()}},
\code{\link[=chk_not_empty]{chk_not_empty()}},
\code{\link[=chk_unique]{chk_unique()}}
}
\concept{misc_checkers}
