% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-atomic.R, R/chk-environment.R
\name{chk_atomic}
\alias{chk_atomic}
\alias{vld_atomic}
\alias{vld_environment}
\title{Check Atomic}
\usage{
chk_atomic(x, x_name = NULL)

vld_atomic(x)

vld_environment(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if atomic using

\code{is.atomic(x)}.
}
\section{Functions}{
\itemize{
\item \code{vld_atomic}: Validate Atomic

\item \code{vld_environment}: Validate Environment
}}

\examples{

# chk_atomic
chk_atomic(1)
try(chk_atomic(list(1)))

# vld_atomic
vld_atomic(1)
vld_atomic(matrix(1:3))
vld_atomic(character(0))
vld_atomic(list(1))
vld_atomic(NULL)

# vld_environment
vld_environment(1)
vld_environment(list(1))
vld_environment(.GlobalEnv)
vld_environment(environment())
}
\seealso{
Other chk_is: \code{\link{chk_environment}}
}
\concept{chk_is}
