% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-character.R
\name{chk_character}
\alias{chk_character}
\alias{vld_character}
\title{Check Character}
\usage{
chk_character(x, x_name = NULL)

vld_character(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails or
returns the original object if successful so it can used in pipes.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if character using

\code{is.character(x)}
}
\section{Functions}{
\itemize{
\item \code{vld_character()}: Validate Character

}}
\examples{
# chk_character
chk_character("1")
try(chk_character(1))
# vld_character
vld_character("1")
vld_character(matrix("a"))
vld_character(character(0))
vld_character(NA_character_)
vld_character(1)
vld_character(TRUE)
vld_character(factor("text"))
}
\seealso{
\code{\link[=is.character]{is.character()}}

For more details about the use of this function,
please read the article
\code{vignette("chk-families")}.

Other data_type_checkers: 
\code{\link{chk_character_or_factor}()},
\code{\link{chk_complex}()},
\code{\link{chk_double}()},
\code{\link{chk_environment}()},
\code{\link{chk_integer}()},
\code{\link{chk_logical}()},
\code{\link{chk_numeric}()},
\code{\link{chk_raw}()}
}
\concept{data_type_checkers}
