% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-complex-number.R
\name{chk_complex_number}
\alias{chk_complex_number}
\alias{vld_complex_number}
\title{Check Complex Number}
\usage{
chk_complex_number(x, x_name = NULL)

vld_complex_number(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails or
returns the original object if successful so it can used in pipes.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if non-missing complex scalar using

\code{is.complex(x) && length(x) == 1L && !anyNA(x)}
}
\section{Functions}{
\itemize{
\item \code{vld_complex_number()}: Validate Complex Number

}}
\examples{
# chk_complex_number
chk_complex_number(as.complex(1.1))
try(chk_complex_number(1.1))
# vld_complex_number
vld_complex_number(as.complex(2))
}
\seealso{
\code{\link[=is.complex]{is.complex()}}

\code{\link[=length]{length()}}

For more details about the use of this function,
please read the article
\href{https://poissonconsulting.github.io/chk/articles/chk-families.html}{chk families}.

Other scalar_checkers:
\code{\link[=chk_count]{chk_count()}},
\code{\link[=chk_date]{chk_date()}},
\code{\link[=chk_date_time]{chk_date_time()}},
\code{\link[=chk_false]{chk_false()}},
\code{\link[=chk_flag]{chk_flag()}},
\code{\link[=chk_lgl]{chk_lgl()}},
\code{\link[=chk_number]{chk_number()}},
\code{\link[=chk_scalar]{chk_scalar()}},
\code{\link[=chk_string]{chk_string()}},
\code{\link[=chk_true]{chk_true()}},
\code{\link[=chk_tz]{chk_tz()}},
\code{\link[=chk_whole_number]{chk_whole_number()}}
}
\concept{scalar_checkers}
