% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-Dates.R
\name{chk_Dates}
\alias{chk_Dates}
\alias{vld_Dates}
\title{Check Date(s)}
\usage{
chk_Dates(x, x_name = NULL)

vld_Dates(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if Date vector using

\code{inherits(x, "Date")}
}
\section{Functions}{
\itemize{
\item \code{vld_Dates}: Validate Date
}}

\examples{
# chk_Dates
chk_Dates(Sys.Date())
try(chk_Dates(1))
# vld_Dates
vld_Dates(Sys.Date())
vld_Dates(1)
}
