% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-environment.R
\name{chk_environment}
\alias{chk_environment}
\title{Check Environment}
\usage{
chk_environment(x, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if environment using

\code{is.environment(x)}.
}
\examples{

# chk_environment
chk_environment(.GlobalEnv)
try(chk_environment(1))
}
\seealso{
Other chk_is: \code{\link{chk_atomic}}
}
\concept{chk_is}
