% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-factor.R
\name{chk_factor}
\alias{chk_factor}
\alias{vld_factor}
\title{Check Factor}
\usage{
chk_factor(x, x_name = NULL)

vld_factor(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails or
returns the original object if successful so it can used in pipes.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if factor using

\code{is.factor(x)}
}
\section{Functions}{
\itemize{
\item \code{vld_factor()}: Validate Factor

}}
\examples{
# chk_factor
chk_factor(factor("1"))
try(chk_factor("1"))
# vld_factor
vld_factor(factor("1"))
vld_factor(factor(0))
vld_factor("1")
vld_factor(1L)
}
\seealso{
\code{\link[=is.factor]{is.factor()}}

For more details about the use of this function,
please read the article
\code{vignette("chk-families")}.

Other factor_checkers:
\code{\link[=chk_character_or_factor]{chk_character_or_factor()}}

Other data_type_checkers:
\code{\link[=chk_character]{chk_character()}},
\code{\link[=chk_character_or_factor]{chk_character_or_factor()}},
\code{\link[=chk_complex]{chk_complex()}},
\code{\link[=chk_double]{chk_double()}},
\code{\link[=chk_environment]{chk_environment()}},
\code{\link[=chk_integer]{chk_integer()}},
\code{\link[=chk_logical]{chk_logical()}},
\code{\link[=chk_number]{chk_number()}},
\code{\link[=chk_numeric]{chk_numeric()}},
\code{\link[=chk_raw]{chk_raw()}}
}
\concept{data_type_checkers}
\concept{factor_checkers}
