% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-file.R, R/deprecated.R
\name{chk_file}
\alias{chk_file}
\alias{vld_file}
\alias{chk_dir}
\alias{vld_dir}
\alias{chk_dirs}
\alias{chk_files}
\title{Check/Validate Files or Directories Exist}
\usage{
chk_file(x)

vld_file(x)

chk_dir(x)

vld_dir(x)

chk_dirs(x)

chk_files(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
The \code{chk_} functions throw an informative error if the test fails.
The \code{vld_} functions return a flag indicating whether the test was met.
}
\description{
Checks/validates if files or directorys exist.
}
\section{Functions}{
\itemize{
\item \code{chk_file}: Check File

Checks if files exist using \code{vld_file()}.

\item \code{vld_file}: Validate File

Checks if files exist using \code{all(\link{file.exists}(x) && !dir.exists(x))}.

\item \code{chk_dir}: Check Directories Exist

Checks if directories exist using \code{vld_dir()}.

\item \code{vld_dir}: Validate Directories Exist

Checks if directories exist using \code{all(\link{dir.exists}(x))}.

\item \code{chk_dirs}: Check Directories Exist (DEPRECATED)

Checks if directories exist using:

\code{all(\link{dir.exists}(x))}

\item \code{chk_files}: Check Files Exist (DEPRECATED)

Checks if files exist using:

\code{all(\link{file.exists}(x) && !dir.exists(x))}
}}

\examples{

# chk_file
try(chk_file(tempfile()))

# vld_file
vld_file(tempfile())

# chk_dir
try(chk_dir(tempfile()))

# vld_dir
vld_dir(tempfile())
}
