% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-length.R, R/chk-standalone.R,
%   R/chk-values.R
\name{chk_length}
\alias{chk_length}
\alias{chk_flag}
\alias{chk}
\alias{chk_lgl}
\alias{chk_number}
\alias{chk_string}
\alias{chk_named}
\alias{chk_null}
\alias{chk_is}
\alias{chk_dim}
\alias{chk_no_missing}
\alias{chk_function}
\alias{chk_unused}
\alias{chk_used}
\alias{chk_dir}
\alias{chk_file}
\alias{chk_values}
\title{Check Object}
\usage{
chk_length(x, length = c(1L, 2147483647), err = TRUE)

chk_flag(x, err = TRUE)

chk_lgl(x, err = TRUE)

chk_number(x, err = TRUE)

chk_string(x, err = TRUE)

chk_named(x, err = TRUE)

chk_null(x, err = TRUE)

chk_is(x, class, err = TRUE)

chk_dim(x, dim, err = TRUE)

chk_no_missing(x, err = TRUE)

chk_function(x, err = TRUE)

chk_unused(..., err = TRUE)

chk_used(..., err = TRUE)

chk_dir(x, err = TRUE)

chk_file(x, err = TRUE)

chk_values(x, values = c(0, Inf, NA), err = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{length}{An integer vector of the permitted length(s). If
two values then it is considered to be a length range.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}

\item{class}{A string specifying the class.}

\item{dim}{An integer vector (or object that can be converted to an integer vector)
of the dimensions of x.}

\item{...}{Additional arguments.}

\item{values}{A vector of the permitted value(s). If
two non-missing values then it is considered to be a value range.}
}
\value{
A flag or an error if the check fails and err == TRUE.
}
\description{
A flag is a non-missing logical scalar.
}
\section{Functions}{
\itemize{
\item \code{chk_length}: Check Length

\item \code{chk_lgl}: Check Logical Scalar

\item \code{chk_number}: Check Number (non-mising numeric scalar)

\item \code{chk_string}: Check String

A string is a non-mising character scalar.

\item \code{chk_named}: Check Named

\item \code{chk_null}: Check NULL

\item \code{chk_is}: Check Inherits from Class

\item \code{chk_dim}: Check Dim

\item \code{chk_no_missing}: Check No Missing Values

\item \code{chk_function}: Check Function

\item \code{chk_unused}: Check ... Unused

\item \code{chk_used}: Check ... Used

\item \code{chk_dir}: Check Directory Exists

\item \code{chk_file}: Check File Exists

A directory is considered to be a file.

\item \code{chk_values}: Check Values

Checks if all x are in values.
}}

\examples{
# chk_flag
chk_flag(1, err = FALSE)
chk_flag(NA, err = FALSE)
chk_flag(TRUE)
chk_flag(c(TRUE, FALSE), err = FALSE)


}
