% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-in.R
\name{chk_in}
\alias{chk_in}
\alias{chk_has}
\title{Check In and Has}
\usage{
chk_in(x, values = c(0, 1, NA), err = TRUE, x_name = NULL)

chk_has(x, values, err = TRUE, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{values}{A vector of the permitted values.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Check in and has.
}
\details{
\code{chk_in()}: Check In

Checks if all values in values using equivalent of:

\code{all(\link{match}(x, values, nomatch = 0) > 0)}
}
\section{Functions}{
\itemize{
\item \code{chk_has}: Check Has Values

Checks if has values using:

\code{all(\link{match}(values, x, nomatch = 0) > 0)}
}}

\examples{

# chk_in
chk_in(numeric(0), 1:10)
chk_in(1, 1:10)
try(chk_in(11, 1:10))
}
