% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-not-subset.R
\name{chk_not_subset}
\alias{chk_not_subset}
\alias{vld_not_subset}
\title{Check Not Subset}
\usage{
chk_not_subset(x, values, x_name = NULL)

vld_not_subset(x, values)
}
\arguments{
\item{x}{The object to check.}

\item{values}{A vector of the permitted values.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails or
returns the original object if successful so it can used in pipes.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if not all values in values using

\code{!any(x \%in\% values) || !length(x)}
}
\section{Functions}{
\itemize{
\item \code{vld_not_subset()}: Validate Not Subset

}}
\examples{
# chk_not_subset
chk_not_subset(11, 1:10)
try(chk_not_subset(1, 1:10))
# vld_not_subset
vld_not_subset(numeric(0), 1:10)
vld_not_subset(1, 1:10)
vld_not_subset(11, 1:10)
}
\seealso{
\code{\link[=any]{any()}}

\code{\link[=length]{length()}}

For more details about the use of this function,
please read the article
\code{vignette("chk-families")}.

Other set_checkers:
\code{\link[=chk_orderset]{chk_orderset()}},
\code{\link[=chk_setequal]{chk_setequal()}},
\code{\link[=chk_subset]{chk_subset()}},
\code{\link[=chk_superset]{chk_superset()}}
}
\concept{set_checkers}
