% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-number.R
\name{chk_number}
\alias{chk_number}
\alias{vld_number}
\title{Check Number}
\usage{
chk_number(x, x_name = NULL)

vld_number(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails or
returns the original object if successful so it can used in pipes.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if non-missing numeric scalar using

\code{is.numeric(x) && length(x) == 1L && !anyNA(x)}

\strong{Pass}: \code{1}, \code{2L}, \code{log(10)}, \code{-Inf}

\strong{Fail}: \code{"a"}, \code{1:3}, \code{NA_real_}
}
\section{Functions}{
\itemize{
\item \code{vld_number()}: Validate Number

}}
\examples{
# chk_number
chk_number(1.1)
try(chk_number(TRUE))
# vld_number
vld_number(1.1)
}
\seealso{
\code{\link[=is.numeric]{is.numeric()}}

\code{\link[=length]{length()}}

For more details about the use of this function,
please read the article
\code{vignette("chk-families")}.

Other data_type_checkers:
\code{\link[=chk_character]{chk_character()}},
\code{\link[=chk_character_or_factor]{chk_character_or_factor()}},
\code{\link[=chk_complex]{chk_complex()}},
\code{\link[=chk_double]{chk_double()}},
\code{\link[=chk_environment]{chk_environment()}},
\code{\link[=chk_factor]{chk_factor()}},
\code{\link[=chk_integer]{chk_integer()}},
\code{\link[=chk_logical]{chk_logical()}},
\code{\link[=chk_numeric]{chk_numeric()}},
\code{\link[=chk_raw]{chk_raw()}}

Other scalar_checkers:
\code{\link[=chk_complex_number]{chk_complex_number()}},
\code{\link[=chk_count]{chk_count()}},
\code{\link[=chk_date]{chk_date()}},
\code{\link[=chk_date_time]{chk_date_time()}},
\code{\link[=chk_false]{chk_false()}},
\code{\link[=chk_flag]{chk_flag()}},
\code{\link[=chk_lgl]{chk_lgl()}},
\code{\link[=chk_scalar]{chk_scalar()}},
\code{\link[=chk_string]{chk_string()}},
\code{\link[=chk_true]{chk_true()}},
\code{\link[=chk_tz]{chk_tz()}},
\code{\link[=chk_whole_number]{chk_whole_number()}}
}
\concept{data_type_checkers}
\concept{scalar_checkers}
