% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on-chk.R
\name{chk_on}
\alias{chk_on}
\alias{chk_off}
\alias{is_chk_on}
\title{Turn checking on and off}
\usage{
chk_on()

chk_off()

is_chk_on()
}
\description{
Turns checking on and off.
}
\section{Functions}{
\itemize{
\item \code{chk_off}: Turns checking off

\item \code{is_chk_on}: Tests checking on
}}

\examples{
print(is_chk_on())
if (is_chk_on()) {
  chk_off()
  print(is_chk_on())
  chk_on()
} else {
  chk_on()
  print(is_chk_on())
  chk_off()
}
}
