% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-r6-class.R
\name{chk_r6_class}
\alias{chk_r6_class}
\alias{vld_r6_class}
\title{Check Inherits from R6 Class}
\usage{
chk_r6_class(x, class, x_name = NULL)

vld_r6_class(x, class)
}
\arguments{
\item{x}{The object to check.}

\item{class}{A character vector specifying the possible class values.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails or
returns the original object if successful so it can used in pipes.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks inherits from R6 class using

\code{inherits(x, "R6") && inherits(x, class)}
}
\details{
All R6 objects inherit from the class \code{'R6'}, so \code{chk_r6_class(x, "R6")}
checks that \code{x} is an R6 object of any class.
R6 class generators (as returned by \code{\link[R6:R6Class]{R6::R6Class()}}) are not themselves
R6 objects.
}
\section{Functions}{
\itemize{
\item \code{vld_r6_class()}: Validate Inherits from R6 Class

}}
\examples{
\dontshow{if (requireNamespace("R6", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# chk_r6_class
chk_r6_class(R6::R6Class("exampleR6class")$new(), "exampleR6class")
try(chk_r6_class(1, "numeric"))
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("R6", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# vld_r6_class
vld_r6_class(R6::R6Class("exampleR6class")$new(), "exampleR6class")
vld_r6_class(1, "numeric")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=inherits]{inherits()}}

For more details about the use of this function,
please read the article
\code{vignette("chk-families")}.

Other id_checkers:
\code{\link[=chk_class]{chk_class()}},
\code{\link[=chk_data]{chk_data()}},
\code{\link[=chk_is]{chk_is()}},
\code{\link[=chk_s3_class]{chk_s3_class()}},
\code{\link[=chk_s3_class_strict]{chk_s3_class_strict()}},
\code{\link[=chk_s4_class]{chk_s4_class()}},
\code{\link[=chk_s7_class]{chk_s7_class()}}
}
\concept{id_checkers}
