% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-s7-class.R
\name{chk_s7_class}
\alias{chk_s7_class}
\alias{vld_s7_class}
\title{Check Inherits from S7 Class}
\usage{
chk_s7_class(x, class, x_name = NULL)

vld_s7_class(x, class)
}
\arguments{
\item{x}{The object to check.}

\item{class}{A character vector specifying the possible class values.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails or
returns the original object if successful so it can used in pipes.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks inherits from S7 class using

\code{inherits(x, "S7_object") && inherits(x, class)}
}
\details{
All S7 objects inherit from the class \code{'S7_object'}, so
\code{chk_s7_class(x, "S7_object")} checks that \code{x} is an S7 object of any class.

S7 is self-describing, so an S7 class generator (as returned by
\code{\link[S7:new_class]{S7::new_class()}}) is itself an S7 object, of class \code{'S7_class'}.
It is not an object of the class it generates.

\code{class} is a character vector of class names matching those in \code{class(x)}.
S7 qualifies class names with the package they are defined in, so an S7
class defined in package \code{foo} is matched by \code{'foo::ClassName'} and not by
\code{'ClassName'}.
S7 class objects are not accepted as \code{class}.
}
\section{Functions}{
\itemize{
\item \code{vld_s7_class()}: Validate Inherits from S7 Class

}}
\examples{
\dontshow{if (requireNamespace("S7", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# chk_s7_class
Foo <- S7::new_class("Foo", package = "mypkg")
chk_s7_class(Foo(), "mypkg::Foo")
try(chk_s7_class(1, "numeric"))
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("S7", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# vld_s7_class
Foo <- S7::new_class("Foo", package = "mypkg")
vld_s7_class(Foo(), "mypkg::Foo")
vld_s7_class(1, "numeric")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=inherits]{inherits()}}

For more details about the use of this function,
please read the article
\code{vignette("chk-families")}.

Other id_checkers:
\code{\link[=chk_class]{chk_class()}},
\code{\link[=chk_data]{chk_data()}},
\code{\link[=chk_is]{chk_is()}},
\code{\link[=chk_r6_class]{chk_r6_class()}},
\code{\link[=chk_s3_class]{chk_s3_class()}},
\code{\link[=chk_s3_class_strict]{chk_s3_class_strict()}},
\code{\link[=chk_s4_class]{chk_s4_class()}}
}
\concept{id_checkers}
