% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-setordered.R
\name{chk_setordered}
\alias{chk_setordered}
\title{Check Set Ordered}
\usage{
chk_setordered(x, values, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{values}{A vector of the permitted values.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if the first occurrence of each shared element
in x is equivalent to the first occurrence of each shared element in values using
\code{vld_equivalent(unique(x[x \%in\% values]), values[values \%in\% x])}.
}
\examples{

# chk_setordered
chk_setordered(1:2, 1:2)
try(chk_setordered(2:1, 1:2))
}
\seealso{
Other chk_set: 
\code{\link{chk_setequal}()},
\code{\link{chk_subset}()},
\code{\link{chk_superset}()}
}
\concept{chk_set}
