% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-string.R
\name{chk_string}
\alias{chk_string}
\alias{chk_match}
\title{Check String}
\usage{
chk_string(x, err = TRUE, x_name = NULL)

chk_match(x, regexp = ".+", err = TRUE, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{err}{Deprecated.}

\item{x_name}{A string of the name of object x or NULL.}

\item{regexp}{A string of a regular expression.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Check if string.
}
\details{
\code{chk_-string()}: Check String

Checks if non-missing character scalar using:

\code{is.character(x) && length(x) == 1L && !anyNA(x)}
}
\section{Functions}{
\itemize{
\item \code{chk_match}: Check Matches

Checks if all values match regular expression using:

\code{all(\link{grepl}(regexp, x))}

\code{regexp} should be a non-missing character scalar.
}}

\examples{

# chk_string
try(chk_string(1))
chk_string("1")
chk_string("")
try(chk_string(NA_character_))
try(chk_string(c("1", "1")))
 

# chk_match
try(chk_match(""))
chk_match("1")
try(chk_match("1", regexp = "2"))
try(chk_match(NA_character_, regexp = ".*"))
chk_match(c("a"), regexp = ".*")
}
