% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-unnamed.R
\name{chk_unnamed}
\alias{chk_unnamed}
\alias{vld_unnamed}
\title{Check Unnamed}
\usage{
chk_unnamed(x, x_name = NULL)

vld_unnamed(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if unnamed using

\code{is.null(names(x))}
}
\section{Functions}{
\itemize{
\item \code{vld_unnamed}: Validate Unnamed
}}

\examples{

# chk_unnamed
chk_unnamed(1)
try(chk_unnamed(c(x = 1)))

# vld_unnamed
vld_unnamed(1)
vld_unnamed(c(x = 1))
}
\concept{chk_miscellaneous}
