% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-unused.R
\name{chk_unused}
\alias{chk_unused}
\alias{chk_used}
\title{Check ... Unused or Used}
\usage{
chk_unused(..., err = TRUE)

chk_used(..., err = TRUE)
}
\arguments{
\item{...}{Additional arguments.}

\item{err}{Deprecated.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Checks if ... is unused or used.
}
\details{
\code{chk_unused()}: Check Unused

Checks if is ... unused using:

\code{!length(\link{list}(...))}
}
\section{Functions}{
\itemize{
\item \code{chk_used}: Check ... Used

Checks if is ... used using:

\code{length(\link{list}(...))}
}}

\examples{

# chk_unused
fun <- function(x, ...) { chk_unused(...); x }
fun(1)
try(fun(1,2))
 

# chk_used
fun <- function(x, ...) { chk_used(...); x }
try(fun(1))
fun(1,2)

}
