% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-used.R
\name{chk_used}
\alias{chk_used}
\alias{vld_used}
\title{Check ... Used}
\usage{
chk_used(...)

vld_used(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
The \code{chk_} function throws an informative error if the test fails.
}
\description{
Checks if is ... used using

\code{length(list(...)) != 0L}
}
\section{Functions}{
\itemize{
\item \code{vld_used()}: Validate ... Used

}}
\examples{
# chk_used
fun <- function(x, ...) {
  chk_used(...)
  x
}
try(fun(1))
fun(1, 2)
# vld_used
fun <- function(x, ...) {
  vld_used(...)
}
fun(1)
fun(1, 2)
}
\seealso{
Other chk_ellipsis: 
\code{\link{chk_unused}()}
}
\concept{chk_ellipsis}
