% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-vector-lengths.R
\name{chk_vector_lengths}
\alias{chk_vector_lengths}
\alias{vld_vector_lengths}
\title{Check Vectors have the Same Length}
\usage{
chk_vector_lengths(...)

vld_vector_lengths(...)
}
\arguments{
\item{...}{The vectors to check the lengths of.}
}
\value{
The \code{chk_} function throws an informative error if the test fails.
}
\description{
To combine vectors they either need to all be the same length or be of length
one. If they are length one then the value will be recycled in the longer
vectors.
}
\section{Functions}{
\itemize{
\item \code{vld_vector_lengths()}: Check Vectors have the Same Length

}}
\examples{
a <- c(1, 2, 3)
b <- c(1, 2, 3)
chk_vector_lengths(a, b)

hour <- c(7, 6, 8)
min <- c(30, 45, 30)
sec <- 0
chk_vector_lengths(hour, min, sec)
a <- c(1, 2, 3)
b <- c(1, 2, 3)
vld_vector_lengths(a, b)

hour <- c(7, 6, 8)
min <- c(30, 45, 30)
sec <- 0
vld_vector_lengths
}
