% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkor.R
\name{chkor}
\alias{chkor}
\title{Check OR}
\usage{
chkor(..., err = TRUE)
}
\arguments{
\item{...}{Multiple \code{\link{chk}} functions.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}
}
\value{
A flag or an error if the check fails and err == TRUE.
}
\description{
Check OR
}
\examples{
chkor()
chkor(chk_flag(TRUE))
chkor(chk_flag(1), err = FALSE)
chkor(chk_flag(1), chk_flag(2), err = FALSE)
chkor(chk_flag(1), chk_flag(TRUE))
}
