% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deparse_tick}
\alias{deparse_tick}
\title{Deparse Tick}
\usage{
deparse_tick(x)
}
\arguments{
\item{x}{A substituted object to deparse.}
}
\value{
A string of the back ticked substituted object.
}
\description{
\code{deparse_tick} is a wrapper on \code{\link{deparse}} that
also back ticks the return string.
}
\details{
It is exported to allow users to construct their own `chk_` functions.
}
\examples{
deparse_tick(2)
deparse_tick(2^2)
}
\seealso{
\code{\link{deparse}}
}
