% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{msg}
\alias{msg}
\title{Message}
\usage{
msg(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or (for
    \code{message} only) a single condition object.}
}
\description{
A wrapper on \code{\link[base]{message}(...)}.
}
\examples{
\dontrun{
msg("this is a message")
}
}
