% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{p}
\alias{p}
\alias{p0}
\title{Concatenate Strings}
\usage{
p(..., sep = " ", collapse = NULL, recycle0 = FALSE)

p0(..., collapse = NULL, recycle0 = FALSE)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{sep}{a character string to separate the terms.  Not
    \code{\link[base]{NA_character_}}.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.  When \code{collapse} is a string,
    the result is always a string (\code{\link[base]{character}} of length 1).}

\item{recycle0}{\code{\link[base]{logical}} indicating if zero-length
    character arguments should result in the zero-length
    \code{\link[base]{character}(0)}.  Note that when \code{collapse} is
    a string, \code{recycle0} does \emph{not} recycle to zero-length, but
    to \code{""}.}
}
\value{
A character vector.
}
\description{
A wrapper on \code{\link[base:paste]{base::paste()}}.
}
\section{Functions}{
\itemize{
\item \code{p0()}: A wrapper on \code{\link[base:paste0]{base::paste0()}}

}}
\examples{
p("a", "b")
p(c("a", "b"), collapse = " ")
p("a", character(0))
p("a", character(0), recycle0 = TRUE)
p0("a", "b")
p0(c("a", "b"), collapse = "")
p0("a", character(0))
p0("a", character(0), recycle0 = TRUE)
}
