% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions}
\usage{
params(x, x_name = NULL, chk = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}

\item{chk}{A flag specifying whether to check the other parameters.}
}
\description{
A non-exported dummy function with key parameter descriptions
that can be inherited by other functions.
}
