% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-package.R
\docType{package}
\name{chk-package}
\alias{chk}
\alias{chk-package}
\title{chk: Check User-Supplied Function Arguments}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

For developers to check user-supplied function
    arguments.  It is designed to be simple, fast and customizable.  Error
    messages follow the tidyverse style guide.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/chk}
  \item Report bugs at \url{https://github.com/poissonconsulting/chk/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

Authors:
\itemize{
  \item Kirill Müller (0000-0002-1416-3412)
}

Other contributors:
\itemize{
  \item Ayla Pearson (0000-0001-7388-1222) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
