% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-package.R
\docType{package}
\name{chk-package}
\alias{chk}
\alias{chk-package}
\title{chk: Check User-Supplied Function Arguments}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

For developers to check user-supplied function arguments.  It
    is designed to be simple, fast and customizable.  Error messages
    follow the tidyverse style guide.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/chk}
  \item Report bugs at \url{https://github.com/poissonconsulting/chk/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Kirill Müller (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
}

Other contributors:
\itemize{
  \item Ayla Pearson (\href{https://orcid.org/0000-0001-7388-1222}{ORCID}) [contributor]
  \item Nadine Hussein \email{nadine@poissonconsulting.ca} (\href{https://orcid.org/0000-0003-4470-8361}{ORCID}) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
