% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-all.R
\name{chk_all}
\alias{chk_all}
\alias{chk_all_identical}
\title{Check All or All Identical}
\usage{
chk_all(x, chk_fun, ..., err = TRUE, x_name = NULL)

chk_all_identical(x, err = TRUE, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{chk_fun}{A chk function.}

\item{...}{Additional arguments.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Checks all elements.
}
\details{
\code{chk_all()}: Check All

Checks all elements using:

\code{all(vapply(x, chk_fun, TRUE, ...))}
}
\section{Functions}{
\itemize{
\item \code{chk_all_identical}: Check All Identical

Checks all elements in x identical using:

\code{all(vapply(x, identical, TRUE, y = x[[1]]))}
}}

\examples{

# chk_all
try(chk_all(1, chk_lgl))
chk_all(TRUE, chk_lgl)
chk_all(c(TRUE, NA), chk_lgl)
chk_all(numeric(0), chk_lgl)
 

# chk_all_identical
try(chk_all_identical(1, 1.0000001))
chk_all_identical(1, 1.00000001)
chk_all_identical(1, 1L)
try(chk_all_identical(c(1,2)))
try(chk_all_identical(c(1,2.0000001)))
 
}
