% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-date.R
\name{chk_date}
\alias{chk_date}
\alias{chk_datetime}
\title{Check Date}
\usage{
chk_date(x, err = TRUE, x_name = NULL)

chk_datetime(x, err = TRUE, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Checks if date or datetime.
}
\details{
\code{chk_date()}: Check Date

Checks if non-missing Date scalar using:

\code{inherits(x, "Date") && length(x) == 1L && !anyNA(x)}
}
\section{Functions}{
\itemize{
\item \code{chk_datetime}: Check DateTime

Checks if non-missing POSIXct scalar using:

\code{inherits(x, "POSIXct") && length(x) == 1L && !anyNA(x)}
}}

\examples{

# chk_date
try(chk_date(1))
try(chk_date("2001-01-02"))
chk_date(as.Date("2001-01-02"))
chk_date(Sys.Date())
try(chk_date(Sys.time()))
try(chk_date(c(Sys.Date(), Sys.Date())))
  

# chk_datetime
try(chk_datetime(1))
try(chk_datetime("2001-01-02"))
chk_datetime(as.POSIXct("2001-01-02"))
try(chk_datetime(Sys.Date()))
chk_datetime(Sys.time())
try(chk_datetime(c(Sys.time(), Sys.time())))
   
}
\seealso{
\code{\link{chk_true}()} and \code{\link{chk_number}()}
}
