% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{chk_deprecated}
\alias{chk_deprecated}
\alias{chk_count}
\alias{chk_proportion}
\title{Deprecated functions}
\usage{
chk_count(x, x_name = NULL)

chk_proportion(x, x_name = NULL)
}
\description{
Deprecated \code{chk_()} functions.
}
\section{Functions}{
\itemize{
\item \code{chk_count}: Check Count
 
Checks if non-missing non-negative integer scalar or double equivalent using:

\code{is.numeric(x) && length(x) == 1L && !anyNA(x) && x >= 0 && 
  isTRUE(all.equal(x, as.integer(x)))}

\item \code{chk_proportion}: Check Proportion

Checks if non-missing numeric scalar between 0 and 1 using:

\code{is.numeric(x) && length(x) == 1L && !anyNA(x) && x >= 0 && x <= 1}
}}

\keyword{internal}
