% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-equal.R
\name{chk_equal}
\alias{chk_equal}
\alias{chk_identical}
\alias{chk_equivalent}
\title{Check Equal, Identical or Equivalent}
\usage{
chk_equal(x, y, tolerance = sqrt(.Machine$double.eps), err = TRUE,
  x_name = NULL)

chk_identical(x, y, err = TRUE, x_name = NULL)

chk_equivalent(x, y, tolerance = sqrt(.Machine$double.eps), err = TRUE,
  x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{y}{An object to check against.}

\item{tolerance}{A non-negative numeric scalar.}

\item{err}{Deprecated.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Checks if two objects are equal, identical or equivalent.
}
\details{
\code{chk_equal()}: Check Equal

Checks if is equal (identical within tolerance) to y using:

\code{isTRUE(\link{all.equal}(x, y, tolerance))}
}
\section{Functions}{
\itemize{
\item \code{chk_identical}: Check Identical

Checks if is identical to y using:

\code{\link{identical}(x, y)}

\item \code{chk_equivalent}: Check Equivalent

Checks if is equivalent (equal ignoring attributes) to y using:

\code{\link{isTRUE}(\link{all.equal}(x, y, tolerance, check.attributes = FALSE))}
}}

\examples{

# chk_equal
try(chk_equal(1, 1.0000001))
chk_equal(1, 1.00000001)
chk_equal(1, 1L)
try(chk_equal(c(x = 1), c(y = 1L)))
chk_equal(c(x = 1), c(x = 1L))
 

# chk_identical
try(chk_identical(1, 1L))
chk_identical(1, 1)
try(chk_identical(1, c(1,1)))
chk_identical(c(1,1), c(1,1))
 

# chk_equivalent
try(chk_equivalent(1, 1.0000001))
chk_equivalent(1, 1.00000001)
chk_equivalent(1, 1L)
chk_equivalent(c(x = 1), c(y = 1L))
chk_equivalent(c(x = 1), c(x = 1L))
 
}
