% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-file.R, R/deprecated.R
\name{chk_file}
\alias{chk_file}
\alias{chk_dir}
\alias{chk_dirs}
\alias{chk_files}
\title{Check Files or Directories Exist}
\usage{
chk_file(x, err = TRUE)

chk_dir(x, err = TRUE)

chk_dirs(x, err = TRUE)

chk_files(x, err = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Checks if files or directorys exist.
}
\details{
\code{chk_file()}: Check Files Exist

Checks if files exist using:

\code{all(\link{file.exists}(x) && !dir.exists(x))}
}
\section{Functions}{
\itemize{
\item \code{chk_dir}: Check Directories Exist

Checks if directories exist using:

\code{all(\link{dir.exists}(x))}

\item \code{chk_dirs}: Check Directories Exist (DEPRECATED)

Checks if directories exist using:

\code{all(\link{dir.exists}(x))}

\item \code{chk_files}: Check Files Exist (DEPRECATED)

Checks if files exist using:

\code{all(\link{file.exists}(x) && !dir.exists(x))}
}}

\examples{

# chk_file
try(chk_file(tempfile()))

# chk_dir
try(chk_dir(tempfile()))
}
