% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-file.R, R/deprecated.R
\name{chk_file}
\alias{chk_file}
\alias{vld_file}
\alias{chk_dir}
\alias{vld_dir}
\alias{chk_ext}
\alias{vld_ext}
\alias{chk_dirs}
\alias{chk_files}
\title{Check/Validate File or Directory Exist}
\usage{
chk_file(x, x_name = NULL)

vld_file(x)

chk_dir(x, x_name = NULL)

vld_dir(x)

chk_ext(x, ext, x_name = NULL)

vld_ext(x, ext)

chk_dirs(x)

chk_files(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}

\item{ext}{A character vector of the permitted file extensions (without the .).}
}
\value{
The \code{chk_} functions throw an informative error if the test fails.
The \code{vld_} functions return a flag indicating whether the test was met.
}
\description{
Checks/validates if file or directory exists.
}
\details{
If checking/validating multiple files use \code{chk_all(x, chk_file)}.
}
\section{Functions}{
\itemize{
\item \code{chk_file}: Check File

Checks if files exist using \code{vld_file(x)}.

\item \code{vld_file}: Validate File

Checks if files exist using \code{all(\link{file.exists}(x) && !dir.exists(x))}.

\item \code{chk_dir}: Check Directories Exist

Checks if directories exist using \code{vld_dir()}.

\item \code{vld_dir}: Validate Directories Exist

Checks if directories exist using \code{vld_string(x) && all(\link{dir.exists}(x))}.

\item \code{chk_ext}: Check File Extension

Checks extension exist using \code{vld_ext(x)}.

The user may want to use \code{\link{toupper}()} or \code{\link{tolower}()}
to ensure the case matches.

\item \code{vld_ext}: Validate File Extension

Checks extension using \code{vld_string(x) && vld_subset(tools::file_ext(x), ext)}.

\item \code{chk_dirs}: Check Directories Exist

\lifecycle{soft-deprecated}

Checks if directories exist using:

\code{all(\link{dir.exists}(x))}

\item \code{chk_files}: Check Files Exist

\lifecycle{soft-deprecated}

Checks if files exist using:

\code{all(\link{file.exists}(x) && !dir.exists(x))}
}}

\examples{

# chk_file
try(chk_file(tempfile()))

# vld_file
vld_file(tempfile())

# chk_dir
try(chk_dir(tempfile()))

# vld_dir
vld_dir(tempfile())

# chk_ext
try(chk_ext("file1.pdf", "png"))

# vld_ext
vld_ext("oeu.pdf", "pdf")
vld_ext(toupper("oeu.pdf"), "PDF")
}
