% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-finite.R
\name{chk_finite}
\alias{chk_finite}
\alias{vld_finite}
\title{Check Finite}
\usage{
chk_finite(x, x_name = NULL)

vld_finite(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails or
returns the original object if successful so it can used in pipes.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if all values are finite using

\code{!is.list(x) && all(is.finite(x))}

Unlike the other range checkers, missing values fail.

\strong{Pass}: \code{1}, \code{c(-1, 0, 1e10)}, \code{TRUE}, \code{numeric(0)}.

\strong{Fail}: \code{NA}, \code{NaN}, \code{Inf}, \code{-Inf}, \code{c(1, NA)}, \code{"1"}, \code{list(1)}.
}
\section{Functions}{
\itemize{
\item \code{vld_finite()}: Validate Finite

}}
\examples{
# chk_finite
chk_finite(1)
try(chk_finite(Inf))
# vld_finite
vld_finite(numeric(0))
vld_finite(c(-1, 0, 1))
vld_finite(c(1, NA))
vld_finite(NaN)
vld_finite(-Inf)
}
\seealso{
\code{\link[=is.finite]{is.finite()}}

\code{\link[=chk_not_any_na]{chk_not_any_na()}}

For more details about the use of this function,
please read the article
\code{vignette("chk-families")}.

Other range_checkers:
\code{\link[=chk_gt]{chk_gt()}},
\code{\link[=chk_gte]{chk_gte()}},
\code{\link[=chk_lt]{chk_lt()}},
\code{\link[=chk_lte]{chk_lte()}},
\code{\link[=chk_range]{chk_range()}}
}
\concept{range_checkers}
