% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-match.R, R/chk-standalone.R
\name{chk_match}
\alias{chk_match}
\alias{chk_flag}
\alias{chk}
\alias{chk_lgl}
\alias{chk_number}
\alias{chk_string}
\alias{chk_named}
\alias{chk_null}
\alias{chk_is}
\alias{chk_identical}
\alias{chk_no_missing}
\alias{chk_function}
\alias{chk_unused}
\alias{chk_used}
\alias{chk_dir}
\alias{chk_file}
\title{Check Object}
\usage{
chk_match(x, match = c(1, Inf), err = TRUE)

chk_flag(x, err = TRUE)

chk_lgl(x, err = TRUE)

chk_number(x, err = TRUE)

chk_string(x, err = TRUE)

chk_named(x, err = TRUE)

chk_null(x, err = TRUE)

chk_is(x, class, err = TRUE)

chk_identical(x, y, err = TRUE)

chk_no_missing(x, err = TRUE)

chk_function(x, err = TRUE)

chk_unused(..., err = TRUE)

chk_used(..., err = TRUE)

chk_dir(x, err = TRUE)

chk_file(x, err = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{match}{A vector of the permitted value(s). If
two non-missing values then it is considered to be a value range.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}

\item{class}{A string specifying the class.}

\item{y}{The object to check against.}

\item{...}{Additional arguments.}
}
\value{
A flag or an error if the check fails and err == TRUE.
}
\description{
A flag is a non-missing logical scalar.
}
\details{
\code{\link{is.logical}(x) && \link{length}(x) == 1L && !\link{is.na}(x)}
}
\section{Functions}{
\itemize{
\item \code{chk_match}: Check Match

\code{all(\link{match}(x, match, nomatch = 0) > 0)}

\item \code{chk_lgl}: Check Logical Scalar

\code{\link{is.logical}(x) && \link{length}(x) == 1L && !\link{is.na}(x)}

\item \code{chk_number}: Check Number (non-mising numeric scalar)

\code{\link{is.numeric}(x) && \link{length}(x) == 1L && !\link{is.na}(x)}

\item \code{chk_string}: Check String

\code{\link{is.character}(x) && \link{length}(x) == 1L && !\link{is.na}(x)}

\item \code{chk_named}: Check Named

\code{!\link{is.null}(\link{names}(x))}

\item \code{chk_null}: Check NULL

\code{\link{is.null}(x)}

\item \code{chk_is}: Check Inherits from Class

\code{\link{inherits}(x, class)}

\item \code{chk_identical}: Check Identical

\code{\link{identical}(x, y)}

\item \code{chk_no_missing}: Check No Missing Values

\code{!\link{anyNA}(x)}

\item \code{chk_function}: Check Function

\code{\link{is.function}(x)}

\item \code{chk_unused}: Check ... Unused

\code{!\link{length}(\link{list}(...))}

\item \code{chk_used}: Check ... Used

\code{\link{length}(\link{list}(...))}

\item \code{chk_dir}: Check Directory Exists

\code{\link{dir.exists}(x)}

\item \code{chk_file}: Check File Exists

\code{\link{file.exists}(x)}
}}

\examples{
# chk_flag
chk_flag(1, err = FALSE)
chk_flag(NA, err = FALSE)
chk_flag(TRUE)
chk_flag(c(TRUE, FALSE), err = FALSE)


}
