% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-in.R
\name{chk_in}
\alias{chk_in}
\alias{vld_in}
\alias{chk_has}
\alias{vld_has}
\title{Check In and Has}
\usage{
chk_in(x, values, x_name = NULL)

vld_in(x, values)

chk_has(x, values, x_name = NULL)

vld_has(x, values)
}
\arguments{
\item{x}{The object to check.}

\item{values}{A vector of the permitted values.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} functions throw an informative error if the test fails.
The \code{vld_} functions return a flag indicating whether the test was met.
}
\description{
Checks/validates if in and has values.
}
\section{Functions}{
\itemize{
\item \code{chk_in}: Check In

Checks if all values in values using \code{vld_in()}.

\item \code{vld_in}: Validate In

Validates all values in values using equivalent of

\code{all(\link{match}(x, values, nomatch = 0) > 0)}

\item \code{chk_has}: Check Has

Checks if includes all values using \code{vld_has()}.

\item \code{vld_has}: Validates Has

Validates includes all values using

\code{all(\link{match}(values, x, nomatch = 0) > 0)}.
}}

\examples{

# chk_in
chk_in(1, 1:10)
try(chk_in(11, 1:10))

# vld_in
vld_in(numeric(0), 1:10)
vld_in(1, 1:10)
vld_in(11, 1:10)

# chk_has
chk_has(1:3, 1)
try(chk_has(1:3, 4))

# vld_has
vld_has(1:3, 1)
vld_has(1:3, 4)
vld_has(integer(0), integer(0))
}
