% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-is.R
\name{chk_is}
\alias{chk_is}
\alias{chk_whole_numeric}
\alias{chk_list}
\alias{chk_function}
\title{Check Is}
\usage{
chk_is(x, class, err = TRUE, x_name = NULL)

chk_whole_numeric(x, err = TRUE, x_name = NULL)

chk_list(x, err = TRUE, x_name = NULL)

chk_function(x, err = TRUE, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{class}{A string specifying the class.}

\item{err}{Deprecated.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Checks if is a particular type of object.
}
\details{
\code{chk_is()}: Check Is

Checks if inherits from class using:

\code{\link{inherits}(x, class)}

Class should be a string.
}
\section{Functions}{
\itemize{
\item \code{chk_whole_numeric}: Check Whole Numeric
 
Checks if integer vector or double equivalent using the equivalent of:

\code{is.integer(x) || (is.double(x) && isTRUE(all.equal(x, as.integer(x))))}

\item \code{chk_list}: Check List

Checks if is a list using:

\code{\link{is.list}(x)}

\item \code{chk_function}: Check Function

Checks if is a function using:

\code{\link{is.function}(x)}
}}

\examples{

# chk_is
try(chk_is(1, "character"))
try(chk_is(NULL, "numeric"))
chk_is(numeric(0), "numeric")
chk_is(1, "numeric")
chk_is(NA_real_, "numeric")
 

# chk_whole_numeric
try(chk_whole_numeric(TRUE))
try(chk_whole_numeric(1.5))
chk_whole_numeric(1)
chk_whole_numeric(NA_real_)
chk_whole_numeric(1:2)
chk_whole_numeric(double(0))
   

# chk_list
try(chk_list(1))
try(chk_list(NULL))
chk_list(list())
chk_list(list(x = 1))
chk_list(mtcars)
 

# chk_function
try(chk_function(1))
try(chk_function(list(1)))
chk_function(mean)
chk_function(function(x) x)
 
}
