% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-is.R
\name{chk_is}
\alias{chk_is}
\alias{vld_is}
\alias{chk_whole_numeric}
\alias{vld_whole_numeric}
\alias{chk_list}
\alias{vld_list}
\alias{chk_function}
\alias{vld_function}
\title{Check/Validate Is}
\usage{
chk_is(x, class, x_name = NULL)

vld_is(x, class)

chk_whole_numeric(x, x_name = NULL)

vld_whole_numeric(x)

chk_list(x, x_name = NULL)

vld_list(x)

chk_function(x, x_name = NULL)

vld_function(x)
}
\arguments{
\item{x}{The object to check.}

\item{class}{A string specifying the class.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} functions throw an informative error if the test fails.
The \code{vld_} functions return a flag indicating whether the test was met.
}
\description{
Checks if is a particular type of object.
}
\section{Functions}{
\itemize{
\item \code{chk_is}: Check Is

Checks if inherits from class using \code{vld_is()}.

\item \code{vld_is}: Validate Is

Validates inherits from class using

\code{\link{inherits}(x, class)}

Class should be a string.

\item \code{chk_whole_numeric}: Check Whole Numeric

Checks if integer vector or double equivalent using \code{vld_whole_numeric()}.

The \code{\link{chk_whole_number}()} function checks if
non-missing integer scalar or double equivalent.

\item \code{vld_whole_numeric}: Validate Whole Numeric

Validates integer vector or double equivalent using

\code{is.integer(x) || (is.double(x) && isTRUE(all.equal(x, as.integer(x))))}

\item \code{chk_list}: Check List

Checks if is a list using
\code{vld_list()}.

\item \code{vld_list}: Validate List

Validates is a list using

\code{\link{is.list}(x)}

\item \code{chk_function}: Check Function

Checks if is a function using \code{vld_function()}.

\item \code{vld_function}: Validate Function

Validates is a function using:

\code{\link{is.function}(x)}
}}

\examples{

# chk_is
chk_is(1, "numeric")
try(chk_is(1, "character"))

# vld_is
vld_is(numeric(0), "numeric")
vld_is(1, "numeric")
vld_is(NA_real_, "numeric")
vld_is(1, "character")
vld_is(NULL, "numeric")

# chk_whole_numeric
chk_whole_numeric(1)
try(chk_whole_numeric(1.1))

# vld_whole_numeric
vld_whole_numeric(1)
vld_whole_numeric(NA_real_)
vld_whole_numeric(1:2)
vld_whole_numeric(double(0))
vld_whole_numeric(TRUE)
vld_whole_numeric(1.5)

# chk_list
chk_list(list())
try(chk_list(1))

# vld_list
vld_list(list())
vld_list(list(x = 1))
vld_list(mtcars)
vld_list(1)
vld_list(NULL)

# chk_function
chk_function(mean)
try(chk_function(1))

# vld_function
vld_function(mean)
vld_function(function(x) x)
vld_function(1)
vld_function(list(1))
}
