% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-null.R
\name{chk_null}
\alias{chk_null}
\alias{chk_not_null}
\title{Check NULL or not NULL}
\usage{
chk_null(x, err = TRUE, x_name = NULL)

chk_not_null(x, err = TRUE, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Checks if NULL or not NULL.
}
\details{
\code{chk_null()}: Check NULL

Checks if NULL using:

\code{\link{is.null}(x)}
}
\section{Functions}{
\itemize{
\item \code{chk_not_null}: Check Not NULL

Checks if not NULL using:

\code{!\link{is.null}(x)}
}}

\examples{

# chk_null
try(chk_null(1))
chk_null(NULL)


# chk_not_null
try(chk_not_null(NULL))
chk_not_null(1)

}
