% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-null.R
\name{chk_null}
\alias{chk_null}
\alias{vld_null}
\title{Check NULL}
\usage{
chk_null(x, x_name = NULL)

vld_null(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails or
returns the original object if successful so it can used in pipes.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if NULL using

\code{is.null(x)}
}
\section{Functions}{
\itemize{
\item \code{vld_null()}: Validate NULL

}}
\examples{
# chk_null
try(chk_null(1))
chk_null(NULL)
# vld_null
vld_null(NULL)
vld_null(1)
}
\seealso{
\code{\link[=is.null]{is.null()}}

For more details about the use of this function,
please read the article
\code{vignette("chk-families")}.

Other null_checkers: 
\code{\link{chk_not_null}()}
}
\concept{null_checkers}
