% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-number.R, R/deprecated.R
\name{chk_number}
\alias{chk_number}
\alias{vld_number}
\alias{chk_whole_number}
\alias{vld_whole_number}
\alias{chk_count}
\alias{chk_proportion}
\title{Check/Validate Number or Whole Number}
\usage{
chk_number(x, x_name = NULL)

vld_number(x)

chk_whole_number(x, x_name = NULL)

vld_whole_number(x)

chk_count(x, x_name = NULL)

chk_proportion(x, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} functions throw an informative error if the test fails.
The \code{vld_} functions return a flag indicating whether the test was met.
}
\description{
Checks/validates if a number or whole number.
}
\section{Functions}{
\itemize{
\item \code{chk_number}: Check Number
 
Checks if non-missing numeric scalar using 
\code{vld_number()}.

\item \code{vld_number}: Validate Number
 
Validates non-missing numeric scalar using

\code{is.numeric(x) && length(x) == 1L && !anyNA(x)}

\item \code{chk_whole_number}: Check Whole Number
 
Checks if non-missing integer scalar or double equivalent using 

\code{vld_whole_number()}.

\item \code{vld_whole_number}: Validate Whole Number
 
Validates non-missing integer scalar or double equivalent using:

\code{is.numeric(x) && length(x) == 1L && !anyNA(x) && 
  (is.integer(x) || isTRUE(all.equal(x, as.integer(x))))}

\item \code{chk_count}: Check Count
 
Checks if non-missing non-negative integer scalar or double equivalent using:

\code{is.numeric(x) && length(x) == 1L && !anyNA(x) && x >= 0 && 
  isTRUE(all.equal(x, as.integer(x)))}

\item \code{chk_proportion}: Check Proportion

Checks if non-missing numeric scalar between 0 and 1 using:

\code{is.numeric(x) && length(x) == 1L && !anyNA(x) && x >= 0 && x <= 1}
}}

\examples{

# chk_number
chk_number(1.1)
try(chk_number(TRUE))

# vld_number
vld_number(1.1)
vld_number(1L)
vld_number(TRUE)
vld_number(NA_real_)
vld_number(c(1.1, 1.1))

# chk_whole_number
chk_whole_number(2)
try(chk_whole_number(1.1))

# vld_whole_number
vld_whole_number(2)
vld_whole_number(2L)
vld_whole_number(NA_integer_)
vld_whole_number(1.1)
vld_whole_number(c(2L, 2L))
}
\seealso{
\code{\link{chk_true}()} and \code{\link{chk_range}()}
}
