% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-number.R, R/deprecated.R
\name{chk_number}
\alias{chk_number}
\alias{chk_whole_number}
\alias{chk_count}
\alias{chk_proportion}
\title{Check Scalar}
\usage{
chk_number(x, err = TRUE, x_name = NULL)

chk_whole_number(x, err = TRUE, x_name = NULL)

chk_count(x, err = TRUE, x_name = NULL)

chk_proportion(x, err = TRUE, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Checks if a non-missing scalar of a specific type.
}
\details{
\code{chk_number()}: Check Number

Checks if non-missing numeric scalar using:

\code{is.numeric(x) && length(x) == 1L && !anyNA(x)}
}
\section{Functions}{
\itemize{
\item \code{chk_whole_number}: Check Whole Number
 
Checks if non-missing integer scalar or double equivalent using:

\code{is.numeric(x) && length(x) == 1L && !anyNA(x) && 
  (is.integer(x) || isTRUE(all.equal(x, as.integer(x))))}

\item \code{chk_count}: Check Count
 
Checks if non-missing non-negative integer scalar or double equivalent using:

\code{is.numeric(x) && length(x) == 1L && !anyNA(x) && x >= 0 && 
  isTRUE(all.equal(x, as.integer(x)))}

\item \code{chk_proportion}: Check Proportion

Checks if non-missing numeric scalar between 0 and 1 using:

\code{is.numeric(x) && length(x) == 1L && !anyNA(x) && x >= 0 && x <= 1}
}}

\examples{

# chk_number
try(chk_number(TRUE))
chk_number(1.1)
chk_number(1L)
try(chk_number(NA_real_))
try(chk_number(c(1.1, 1.1)))


# chk_whole_number
try(chk_whole_number(1.1))
chk_whole_number(2)
chk_whole_number(2L)
try(chk_whole_number(NA_integer_))
try(chk_whole_number(c(2L, 2L)))
  
}
\seealso{
\code{\link{chk_true}()} and \code{\link{chk_range}()}
}
