% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-not-subset.R, R/chk-subset.R
\name{check_not_subset}
\alias{check_not_subset}
\alias{vld_not_subset}
\alias{chk_subset}
\alias{check_subset}
\alias{vld_subset}
\title{Check Subset}
\usage{
check_not_subset(x, values, x_name = NULL)

vld_not_subset(x, values)

chk_subset(x, values, x_name = NULL)

check_subset(x, values, x_name = NULL)

vld_subset(x, values)
}
\arguments{
\item{x}{The object to check.}

\item{values}{A vector of the permitted values.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} function throws an informative error if the test fails.

The \code{check_} function throws an informative error if the test fails and
returns the original object if successful so it can used in pipes.

The \code{vld_} function returns a flag indicating whether the test was met.
}
\description{
Checks if all values in values using

\code{all(x \%in\% values)}
}
\section{Functions}{
\itemize{
\item \code{check_not_subset}: Check Not Subset

\item \code{vld_not_subset}: Validate Not Subset

\item \code{check_subset}: Check Subset

\item \code{vld_subset}: Validate Subset
}}

\examples{
# check_not_subset
check_not_subset(11, 1:10)
try(check_not_subset(1, 1:10))
# vld_not_subset
vld_not_subset(numeric(0), 1:10)
vld_not_subset(1, 1:10)
vld_not_subset(11, 1:10)
# chk_subset
chk_subset(1, 1:10)
try(chk_subset(11, 1:10))
# check_subset
check_subset(1, 1:10)
try(check_subset(11, 1:10))
# vld_subset
vld_subset(numeric(0), 1:10)
vld_subset(1, 1:10)
vld_subset(11, 1:10)
}
\seealso{
Other chk_set: 
\code{\link{check_orderset}()},
\code{\link{chk_join}()},
\code{\link{chk_not_subset}()},
\code{\link{chk_orderset}()},
\code{\link{chk_superset}()}
}
\concept{chk_set}
