% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-true.R
\name{chk_true}
\alias{chk_true}
\alias{vld_true}
\alias{chk_false}
\alias{vld_false}
\title{Check/Validate TRUE, FALSE, Flag or Logical Scalar}
\usage{
chk_true(x, x_name = NULL)

vld_true(x)

chk_false(x, x_name = NULL)

vld_false(x)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
The \code{chk_} functions throw an informative error if the test fails.
The \code{vld_} functions return a flag indicating whether the test was met.
}
\description{
Checks/validates if TRUE, FALSE, a flag (TRUE or FALSE)
or a lgl (TRUE, FALSE, or NA).
}
\section{Functions}{
\itemize{
\item \code{chk_true}: Check TRUE

Checks if TRUE using
\code{vld_true(x)}.

\item \code{vld_true}: Validate TRUE

Validates if TRUE using

\code{identical(unname(x), TRUE)}.

\item \code{chk_false}: Check FALSE

Checks if FALSE using
\code{vld_false(x)}.

\item \code{vld_false}: Validate FALSE

Validates if FALSE using

\code{identical(unname(x), FALSE)}.
}}

\examples{

# chk_true
chk_true(TRUE)
try(chk_true(1))

# vld_true
vld_true(TRUE)
vld_true(FALSE)
vld_true(NA)
vld_false(0)
vld_true(c(TRUE, TRUE))

# chk_false
chk_false(FALSE)
try(chk_false(0))

# vld_false
vld_false(TRUE)
vld_false(FALSE)
vld_false(NA)
vld_false(0)
vld_false(c(FALSE, FALSE))
}
