% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-true.R, R/deprecated.R
\name{chk_true}
\alias{chk_true}
\alias{chk_false}
\alias{chk_flag}
\alias{chk_lgl}
\alias{chk_dirs}
\alias{chk_files}
\title{Check TRUE or FALSE}
\usage{
chk_true(x, err = TRUE, x_name = NULL)

chk_false(x, err = TRUE, x_name = NULL)

chk_flag(x, err = TRUE, x_name = NULL)

chk_lgl(x, err = TRUE, x_name = NULL)

chk_dirs(x, err = TRUE)

chk_files(x, err = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Checks if object is TRUE or FALSE.
}
\details{
\code{chk_true()}: Check TRUE

Checks if TRUE using:

\code{\link{isTRUE}(x)}
}
\section{Functions}{
\itemize{
\item \code{chk_false}: Check FALSE

Checks if FALSE using:

\code{\link{isFALSE}(x)}

\item \code{chk_flag}: Check Flag

Checks if non-missing logical scalar using:

\code{is.logical(x) && length(x) == 1L && !anyNA(x)}

\item \code{chk_lgl}: Check Lgl

Checks if logical scalar using:

\code{is.logical(x) && length(x) == 1L && !anyNA(x)}

\item \code{chk_dirs}: Check Directories Exist (DEPRECATED)

Checks if directories exist using:

\code{all(\link{dir.exists}(x))}

\item \code{chk_files}: Check Files Exist (DEPRECATED)

Checks if files exist using:

\code{all(\link{file.exists}(x) && !dir.exists(x))}
}}

\examples{

# chk_true
try(chk_true(1))
try(chk_true(NA))
try(chk_true(FALSE))
chk_true(TRUE)
try(chk_true(c(TRUE, TRUE)))

# chk_false
try(chk_false(0))
try(chk_false(NA))
try(chk_false(TRUE))
chk_false(FALSE)
try(chk_false(c(FALSE, FALSE)))


# chk_flag
try(chk_flag(1))
try(chk_flag(NA))
chk_flag(TRUE)
chk_flag(FALSE)
try(chk_flag(c(TRUE, TRUE)))


# chk_lgl
try(chk_lgl(1))
chk_lgl(NA)
chk_lgl(TRUE)
chk_lgl(FALSE)
try(chk_lgl(c(TRUE, TRUE)))

}
\seealso{
\code{\link{chk_number}()}
}
