% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-unique.R, R/deprecated.R
\name{chk_unique}
\alias{chk_unique}
\alias{chk_no_missing}
\alias{chk_named}
\alias{chk_in}
\alias{chk_match}
\alias{chk_length}
\title{Check Unique}
\usage{
chk_unique(x, incomparables = FALSE, err = TRUE, x_name = NULL)

chk_no_missing(x, err = TRUE, x_name = NULL)

chk_named(x, err = TRUE, x_name = NULL)

chk_in(x, values = c(0, 1, NA), err = TRUE, x_name = NULL)

chk_match(x, regexp = ".+", err = TRUE, x_name = NULL)

chk_length(x, length = 1L, err = TRUE, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{incomparables}{A vector of values that cannot be compared.
FALSE means that all values can be compared.}

\item{err}{A flag specifying whether to generate an error
message if the check fails.}

\item{x_name}{A string of the name of object x or NULL.}

\item{values}{A vector of the permitted values.}

\item{regexp}{A string of a regular expression.}

\item{length}{A count of the length.}
}
\value{
TRUE if passes check. Otherwise if throws an informative error unless
\code{err = FALSE} in which case it returns FALSE.
}
\description{
Unique checks.
}
\details{
\code{chk_unique()}: Check Unique

Checks if unique using:

\code{!\link{anyDuplicated}(x, incomparables = incomparables)}
}
\section{Functions}{
\itemize{
\item \code{chk_no_missing}: Check No Missing Values

Checks if no missing values using:

\code{!\link{anyNA}(x)}

\item \code{chk_named}: Check Named

Checks if is named using:

\code{!is.null(\link{names}(x))}

\item \code{chk_in}: Check In

Checks if all values in values using equivalent of:

\code{all(\link{match}(x, values, nomatch = 0) > 0)}

\item \code{chk_match}: Check Matches

Checks if all values match regular expression using:

\code{all(\link{grepl}(regexp, x))}

\code{regexp} should be a non-missing character scalar.

\item \code{chk_length}: Check Length


Checks if is length length using:

\code{length(x) == length}

\code{length} should be a count.
}}

\examples{

# chk_unique
chk_unique(NULL)
chk_unique(numeric(0))
chk_unique(c(NA, 2))
try(chk_unique(c(NA, NA, 2)))
chk_unique(c(NA, NA, 2), incomparables = NA)


# chk_no_missing
try(chk_no_missing(NA))
try(chk_no_missing(c(1, NA)))
chk_no_missing(character(0))
chk_no_missing(1)
chk_no_missing(1:10)
  

# chk_named
try(chk_named(1))
try(chk_named(list(1)))
chk_named(c(x = 1))
chk_named(list(x = 1))
chk_named(c(x = 1)[-1])
chk_named(list(x = 1)[-1])
 

# chk_in
chk_in(numeric(0), 1:10)
chk_in(1, 1:10)
try(chk_in(11, 1:10))

# chk_match
try(chk_match(""))
chk_match("1")
try(chk_match("1", regexp = "2"))
try(chk_match(NA_character_, regexp = ".*"))
chk_match(c("a"), regexp = ".*")
}
