% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-unused.R
\name{chk_unused}
\alias{chk_unused}
\alias{vld_unused}
\title{Check ... Unused}
\usage{
chk_unused(...)

vld_unused(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
The \code{chk_} function throws an informative error if the test fails.
}
\description{
Checks if ... is unused

\code{length(list(...)) == 0L}
}
\section{Functions}{
\itemize{
\item \code{vld_unused()}: Validate ... Unused

}}
\examples{
# chk_unused
fun <- function(x, ...) {
  chk_unused(...)
  x
}
fun(1)
try(fun(1, 2))
# vld_unused
fun <- function(x, ...) {
  vld_unused(...)
}
fun(1)
try(fun(1, 2))
}
\seealso{
Other chk_ellipsis: 
\code{\link{chk_used}()}
}
\concept{chk_ellipsis}
