% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkor.R
\name{chkor}
\alias{chkor}
\title{Check OR}
\usage{
chkor(...)
}
\arguments{
\item{...}{Multiple \code{chk_} functions.}
}
\value{
An informative error if the test fails.
}
\description{
Check OR
}
\examples{
chkor()
chkor(chk_flag(TRUE))
try(chkor(chk_flag(1)))
try(chkor(chk_flag(1), chk_flag(2)))
chkor(chk_flag(1), chk_flag(TRUE))
}
