% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkor.R
\name{chkor}
\alias{chkor}
\title{Check OR}
\usage{
chkor(...)
}
\arguments{
\item{...}{Multiple \code{chk_} functions.}
}
\value{
An informative error if the test fails.
}
\description{
The \code{chkor()} function has been deprecated for the faster \code{chkor_vld()}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
chkor()
chkor(chk_flag(TRUE))
try(chkor(chk_flag(1)))
try(chkor(chk_flag(1), chk_flag(2)))
chkor(chk_flag(1), chk_flag(TRUE))
}
\seealso{
\code{\link[=chk_null_or]{chk_null_or()}}
}
