% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/utils.R
\name{deparse_backtick}
\alias{deparse_backtick}
\alias{deparse_backtick_chk}
\alias{backtick_chk}
\title{Deparse Back Tick}
\usage{
deparse_backtick(x)

deparse_backtick_chk(x)

backtick_chk(x)
}
\arguments{
\item{x}{A substituted object to deparse.}
}
\value{
A string of the back ticked substituted object.
}
\description{
\code{deparse_backtick_chk} is a wrapper on \code{\link{deparse}} 
and \code{backtick_chk}.
}
\details{
It is exported to allow users to easily construct their own `chk_` functions.
}
\section{Functions}{
\itemize{
\item \code{deparse_backtick}: Deparse Back Tick

\lifecycle{soft-deprecated}

\item \code{backtick_chk}: Back Tick
}}

\examples{

# deparse_backtick_chk
deparse_backtick_chk(2)
deparse_backtick_chk(2^2)
}
\seealso{
\code{\link{deparse}}
}
