% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deparse_backtick_chk}
\alias{deparse_backtick_chk}
\alias{backtick_chk}
\alias{unbacktick_chk}
\title{Deparse Backtick}
\usage{
deparse_backtick_chk(x)

backtick_chk(x)

unbacktick_chk(x)
}
\arguments{
\item{x}{A substituted object to deparse.}
}
\value{
A string of the backticked substituted object.
}
\description{
\code{deparse_backtick_chk} is a wrapper on \code{\link[=deparse]{deparse()}}
and \code{backtick_chk}.
}
\details{
It is exported to allow users to easily construct their own \code{chk_} functions.
}
\section{Functions}{
\itemize{
\item \code{backtick_chk}: Backtick

\item \code{unbacktick_chk}: Unbacktick
}}

\examples{

# deparse_backtick_chk
deparse_backtick_chk(2)
deparse_backtick_chk(2^2)
}
\seealso{
\code{\link[=deparse]{deparse()}}
}
