% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-chk.R
\name{err}
\alias{err}
\alias{wrn}
\alias{msg}
\title{Stop, Warning and Message}
\usage{
err(..., n = NULL, .subclass = NULL)

wrn(..., n = NULL, .subclass = NULL)

msg(..., n = NULL, .subclass = NULL)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}

\item{n}{The value of n for converting \code{sprintf}-like types.}

\item{.subclass}{Subclass of the condition. This allows your users
to selectively handle the conditions signalled by your functions.}
}
\description{
Wrappers on 
\code{\link[rlang]{abort}()}, \code{\link[rlang]{warn}()} and
 \code{\link[rlang]{inform}()}.
}
\section{Functions}{
\itemize{
\item \code{err}: Error

\item \code{wrn}: Warning

\item \code{msg}: Message
}}

\section{\code{sprintf}-like types}{

The following \code{sprintf}-like types can be used in the custom messages:

\describe{
  \item{\code{n}}{The value of n.}
  \item{\code{s}}{'' if n == 1 otherwise 's'}
  \item{\code{r}}{'is' if n == 1 otherwise 'are'}
}
}

\examples{

# err
try(err("this is an error"))

# wrn
wrn("this is a warning")
wrn("there \%r \%n problem value\%s")
wrn("there \%r \%n problem value\%s", n = 1)
wrn("there \%r \%n problem value\%s", n = 2)

# msg
msg("this is a message")
msg("there \%r \%n problem value\%s")
msg("there \%r \%n problem value\%s", n = 1)
msg("there \%r \%n problem value\%s", n = 2)
}
