% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{err}
\alias{err}
\title{Stop}
\usage{
err(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}
}
\description{
A wrapper on \code{\link[base]{stop}(..., call. = FALSE)}.
}
\examples{
try(err("this is an error"))
}
