% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for chk Functions}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}

\item{chk}{A flag specifying whether to check the other parameters.}
}
\description{
Parameter Descriptions for chk Functions
}
\keyword{internal}
