% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for chk Functions}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}

\item{chk}{A flag specifying whether to check the other parameters.}

\item{chk_fun}{A chk_ function.}

\item{tolerance}{A non-negative numeric scalar.}

\item{ext}{A character vector of the permitted file extensions (without the .).}

\item{y}{An object to check against.}

\item{value}{A non-missing scalar of a value.}

\item{range}{A non-missing sorted vector of length 2 of the lower and upper permitted values.}

\item{regexp}{A string of a regular expression.}

\item{values}{A vector of the permitted values.}

\item{class}{A string specifying the class.}

\item{formals}{A count of the number of formal arguments.}

\item{incomparables}{A vector of values that cannot be compared.
FALSE means that all values can be compared.}

\item{...}{Additional arguments.}
}
\description{
Parameter Descriptions for chk Functions
}
\keyword{internal}
