% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-all.R, R/chk-identical.R, R/chk-number.R,
%   R/vld.R
\name{vld_all}
\alias{vld_all}
\alias{vld_all_identical}
\alias{vld_all_equal}
\alias{vld_all_equivalent}
\alias{vld_identical}
\alias{vld_equal}
\alias{vld_equivalent}
\alias{vld_number}
\alias{vld_whole_number}
\alias{vld}
\title{Validators}
\usage{
vld_all(x, vld_fun, ...)

vld_all_identical(x)

vld_all_equal(x, tolerance = sqrt(.Machine$double.eps))

vld_all_equivalent(x, tolerance = sqrt(.Machine$double.eps))

vld_identical(x, y)

vld_equal(x, y, tolerance = sqrt(.Machine$double.eps))

vld_equivalent(x, y, tolerance = sqrt(.Machine$double.eps))

vld_number(x)

vld_whole_number(x)
}
\arguments{
\item{x}{The object to check.}

\item{vld_fun}{A vld_ function.}

\item{...}{Additional arguments.}

\item{tolerance}{A non-negative numeric scalar.}

\item{y}{An object to check against.}
}
\value{
A flag indicating whether the object passed the test.
}
\description{
Each \code{chk_()} function has a corresponding \code{vld_()} function.
}
\examples{
vld_all(c(TRUE, NA), vld_lgl)
vld_all_identical(c(1, 1))
vld_all_equal(c(1, 1L))
vld_all_equivalent(c(x = 1, y = 1))
vld_identical(1, 1)
vld_equal(1, 1.00000001)
vld_equivalent(c(x = 1), c(y = 1L)) 
vld_number(1.1)
vld_whole_number(2)
}
