% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-number.R, R/vld.R
\name{vld_number}
\alias{vld_number}
\alias{vld_whole_number}
\alias{vld}
\title{Validators}
\usage{
vld_number(x)

vld_whole_number(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A flag indicating whether the object passed the test.
}
\description{
Each \code{chk_()} function has a corresponding \code{vld_()} function.
}
\examples{
vld_number(1.1)
vld_whole_number(2)
}
