% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld}
\alias{vld}
\title{Validators}
\arguments{
\item{x}{The object to check.}

\item{y}{An object to check against.}

\item{vld_fun}{A vld_ function.}

\item{tolerance}{A non-negative numeric scalar.}

\item{...}{Additional arguments.}
}
\value{
A flag indicating whether the object passed the test.
}
\description{
Each \code{chk_()} function has a corresponding \code{vld_()} function.
}
