% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{wrn}
\alias{wrn}
\title{Warning}
\usage{
wrn(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}
}
\description{
A wrapper on \code{\link[base]{warning}(..., call. = FALSE)}.
}
\examples{
\dontrun{
wrn("this is a warning")
}
}
